and another one
svn:r2484

+++ src/common/util.h
@@ -208,7 +208,7 @@ file_status_t file_status(const char *filename);
 int check_private_dir(const char *dirname, int create);
 int write_str_to_file(const char *fname, const char *str, int bin);
 char *read_file_to_str(const char *filename, int bin);
-int parse_line_from_file(char *line, int maxlen, FILE *f, char **key_out, char **value_out);
+int parse_line_from_file(char *line, size_t maxlen, FILE *f, char **key_out, char **value_out);
 char *expand_filename(const char *filename);
 int replace_file(const char *from, const char *to);
